+++ lib/src/utils/create_spry.dart
@@ -3,10 +3,10 @@ import 'package:routingkit/routingkit.dart';
 import '../types.dart';
 
 /// Creates a new [Spry] application.
-Spry createSpry({Router<Handler>? router, Iterable<Handler>? stack}) {
+Spry createSpry({RouterContext<Handler>? router, Iterable<Handler>? stack}) {
   return _SpryImpl(
     router: switch (router) {
-      Router<Handler> router => router,
+      RouterContext<Handler> router => router,
       _ => createRouter<Handler>(),
     },
     stack: [...?stack],
@@ -17,14 +17,14 @@ class _SpryImpl implements Spry {
   const _SpryImpl({required this.router, required this.stack});
 
   @override
-  final Router<Handler> router;
+  final RouterContext<Handler> router;
 
   @override
   final List<Handler> stack;
 
   @override
   void on<T>(String? method, String path, Handler<T> handler) {
-    router.add(method, path, handler);
+    addRoute(router, method ?? '', path, handler);
   }
 
   @override