+++ CHANGELOG.md
@@ -1,7 +1,11 @@
+## 5.1.0
+
+Adapt to `routingkit` v4.0.0 version
+
 ## 5.0.2
 
-* Remove `http_parser` package.
+- Remove `http_parser` package.
 
 ## 5.0.1
 
-* Change the `meta` package version to `>=1.12.0 <2.0.0`
+- Change the `meta` package version to `>=1.12.0 <2.0.0`